requirements.txt
numpy==1.26.4
opencv-python==4.6.0.66
paddlepaddle==2.6.2
paddleocr==2.7.3
shapely
pyclipper
Pillow
fastapi>=0.111.0,<1.0.0
uvicorn[standard]>=0.30.0,<1.0.0
python-multipart>=0.0.9
pydantic>=2.7.0,<3.0.0
streamlit>=1.34.0,<2.0.0
requests>=2.31.0,<3.0.0
pytest>=8.0.0,<9.0.0
